+++ docker/Dockerfile.gpu
@@ -51,5 +51,9 @@ EXPOSE 7860
 ENV GRADIO_SERVER_NAME=0.0.0.0 \
     GRADIO_SERVER_PORT=7860 \
     GRADIO_SHARE=0
+
+HEALTHCHECK --interval=30s --timeout=10s --start-period=60s \
+  CMD curl -f http://localhost:7860/ || exit 1
+
 CMD ["uv", "run", "vieneu-web"]
