+++ pyproject.toml
@@ -73,6 +73,7 @@ dependencies = [
     "perth>=0.2.0",
     "librosa>=0.11.0",
     "tqdm",
+    "cachetools",
 ]
 
 [project.urls]
